requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# Document Parsers
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# AI APIs
openai>=1.0.0

# Keyword Extraction
yake>=0.4.8

# Google Sheets
gspread>=5.0.0
oauth2client>=4.1.3

# Configuration
python-dotenv>=1.0.0

# Testing
hypothesis>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
chardet>=5.0.0
requests>=2.31.0
